Fire VPC and Firewall branch submit and errors correctly
